Implementing authorization using gates

+++ app/Http/Controllers/QuestionsController.php
@@ -62,6 +62,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access Denied');
+        }
         return view("questions.edit", compact('question'));
     }
 
@@ -74,6 +77,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, 'Access Denied');
+        }
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', "Your question has been updated.");
         // return redirect('/questions')with('success', "Your question has been updated.");
@@ -87,6 +93,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, 'Access Denied');
+        }
         $question->delete();
         return redirect('/questions')->with('success', "Your question has been deleted");
     }
